requirements.txt
tqdm
requests
urllib3
beautifulsoup4
Pillow
fpdf